Update flakes
Now hyprland won't build with the older version of nixpkgs (it appears
to only ever been built with the latest nixpkgs-unstable). The reason I
had pinned it in the first place was because Kitty stopped working
(presumably due to a mismatch with the mesa versions of nixpkgs and
nixpkgs of hyprland). So it makes sense to me to peg kitty directly to
the hyprland nixpkgs, even though that's just the unstable packages
which I could also access - they are coupled so it's good for it to be
explicit.